Build/Edge-node/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
requests
psutil
